+++ agb/src/syscall.rs
@@ -56,7 +56,7 @@ pub fn wait_for_interrupt() {
 
 /// The vblank interrupt handler [VBlank][crate::interrupt::VBlank] should be
 /// used instead of calling this function directly.
-pub fn wait_for_vblank() {
+pub(crate) fn wait_for_vblank() {
     unsafe {
         asm!(
             "swi {SWI}",