+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Text renderer can now be re-used which is useful for rpg style character/word at a time text boxes.
+- If a vblank happens outside of `wait_for_vblank`, then next call will immediately return.
 
 ### Fixed
 - Zero volume now plays no sound.